feat: Added Bike spawning at corners
Every 10 episodes, spawn bikes at all empty corners of the map.

Since the agents don't yet have any programming, it doesn't include any
destinations. Only spawning. But in the future, the Bikes should accept
a random destination on the map to go to, and be able to get to it no
matter their spawning location.

There is no longer a requirement to utilize the number of agents
variable as something that is coupled with the model itself.