refactor: Convert `init` endpoint to GET
The model initialization endpoint no longer needs to receive any sort of
information from the visualization server because the visualization
server doesn't dictate any parameter of the simulation.

An is instead recipient of the information the model initialization yields.
Specifically referring to the width and height of the map. Which is
dynamically read from a `.txt` file and is determined once the model is
initialized.

The endpoint and request have been modified for this purpose to instead
transmit the width and height of the map from the model to the
visualization server.

And the `number_agents` and `NAgents` variables are irrelevant since the
model no longer requires it for initialization either.