fix: Correct file structure for access
Since the `python` process used to run the server is being run in the
`Server/agentsServer` directory, the
`Server/agentsServer/parkAgents/park_files` were unreachable with
the relative paths that were coded in the
`Server/agentsServer/parkAgents/model.py` file. Hence moving them to the
same directory as the one where the `python` process was running was
paramount.