fix: use `result` instead of `response` for `data`
The variable `response` is the raw string body of the response. While
the variable `result` is the JSON-parsed response. Hence it does contain
the desired `height and width values required for the correct
calculation of the viewProjectionMatrix.

This error meant that the `data` object attributes `width` and `height`
remained undefined and corrupted all of the matrices data to NaNs.