feat: convert function export to variable export
Since all the function was doing was returning the attributes of the
object, accessed by a key, then it's best to simply have direct access
to the object attributes. And it requires less function calls and
instead and requires no different information as having to call the
function with the key to access the attribute.