feat: bikes aligned with the road they are on
The bikes now rotate based on their direction, dictated by the direction
of the road beneath them. The wheels also rotate accordingly to align
themselves with the frame of the bike, as well as have their own extra
rotation around the x axis to be able to spin freely. Beyond any
rotations the bike frame undergoes.