fix: model stops and servers stops updating
The model now properly stops when there are no more bikes to spawn and
the visualization server stops asking for further updates. This results
in the visualization server still being able to render the view, but no
longer updating the positions of the agents. This also has the
side-effect that the bikes fly off into infinity because the frameCount
is not reset to 0. But I take it as a funny ending animation.